KAFKA-3431: Move `BrokerEndPoint` from `o.a.k.common` to `o.a.k.common.internals`
