+++ clients/src/main/java/org/apache/kafka/common/internals/BrokerEndPoint.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common;
+package org.apache.kafka.common.internals;
 
 import java.io.Serializable;
